requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.21.1
langchain>=0.1.10
langchain-google-genai>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pillow>=9.0.0
httpx>=0.24.0
markdown-pdf>=0.1.3
markdown>=3.4.3
PyMuPDF>=1.22.5
PyPDF2>=3.0.0
